requirements.txt
blake2b-py
coloredlogs>=10.0
coincurve>=13.0.0
cython
cytoolz>=0.12.0
asn1crypto>=0.22.0
configparser>=3.5.0
coverage<7.0,>6.0
py_ecc<5.0.0,>=1.4.7
eth_abi==5.1.0
eth-account<0.6.0,>=0.5.6
ethereum-input-decoder>=0.2.2
eth-hash<0.4.0,>=0.3.1
eth-keyfile<0.6.0,>=0.5.1
eth-keys<0.4.0,>=0.2.1
eth-rlp<0.3.0,>=0.1.0
eth-typing<3.0.0,>=2.1.0
eth-utils<2
hexbytes<0.3.0
jinja2>=2.9
MarkupSafe<2.1.0
mock
mypy-extensions<1.0.0
persistent>=4.2.0
py-flags
py-evm==0.5.0a1
py-solc-x
py-solc
pytest>=3.6.0
pyparsing<3,>=2.0.2
pytest-cov
pytest_mock
requests
rlp<3
semantic_version
transaction>=2.2.1
typing-extensions<4,>=3.7.4
z3-solver>=4.8.8.0
matplotlib
pre-commit<2.21.0
certifi>=2020.06.20
